pyproject.toml
[project]
name = "ford-desafio-dev-backend"
version = "0.1.0"
description = "Desafio Dev Backend Ford usando FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "alembic>=1.15.1",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[tool.ruff.lint]
extend-select = ["E", "W"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
    "integration",
]
